backend: adapt code after splitting ACL filter into 2 messages
